+++ dotnet-interactive.Tests/CommandLine/TelemetryTests.cs
@@ -5,7 +5,6 @@ using System;
 using System.CommandLine;
 using System.CommandLine.Invocation;
 using System.IO;
-using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.DotNet.Interactive.App.CommandLine;
@@ -310,44 +309,6 @@ namespace Microsoft.DotNet.Interactive.App.Tests.CommandLine
         }
 
         [Fact]
-        public void Telemetry_common_properties_should_contain_if_it_is_in_docker_or_not()
-        {
-            var unitUnderTest = new TelemetryCommonProperties(
-                "product-version",
-                userLevelCacheWriter: new NothingUserLevelCacheWriter());
-            unitUnderTest.GetTelemetryCommonProperties().Should().ContainKey("Docker Container");
-        }
-
-        [Fact]
-        public void Telemetry_common_properties_should_return_hashed_machine_id()
-        {
-            var unitUnderTest = new TelemetryCommonProperties(
-                "product-version",
-                getMACAddress: () => "plaintext", userLevelCacheWriter: new NothingUserLevelCacheWriter());
-            unitUnderTest.GetTelemetryCommonProperties()["Machine ID"].Should().NotBe("plaintext");
-        }
-
-        [Fact]
-        public void Telemetry_common_properties_should_return_new_guid_when_cannot_get_mac_address()
-        {
-            var unitUnderTest = new TelemetryCommonProperties(
-                "product-version",
-                getMACAddress: () => null, userLevelCacheWriter: new NothingUserLevelCacheWriter());
-            var assignedMachineId = unitUnderTest.GetTelemetryCommonProperties()["Machine ID"];
-
-            Guid.TryParse(assignedMachineId, out var _).Should().BeTrue("it should be a guid");
-        }
-
-        [Fact]
-        public void Telemetry_common_properties_should_contain_kernel_version()
-        {
-            var unitUnderTest = new TelemetryCommonProperties(
-                "product-version",
-                getMACAddress: () => null, userLevelCacheWriter: new NothingUserLevelCacheWriter());
-            unitUnderTest.GetTelemetryCommonProperties()["Kernel Version"].Should().Be(RuntimeInformation.OSDescription);
-        }
-
-        [Fact]
         public async Task Show_first_time_message_if_environment_variable_is_not_set()
         {
             var environmentVariableName = FirstTimeUseNoticeSentinel.SkipFirstTimeExperienceEnvironmentVariableName;
@@ -355,7 +316,7 @@ namespace Microsoft.DotNet.Interactive.App.Tests.CommandLine
             Environment.SetEnvironmentVariable(environmentVariableName, null);
             try
             {
-                await _parser.InvokeAsync(String.Empty, _console);
+                await _parser.InvokeAsync("jupyter", _console);
                 _console.Out.ToString().Should().Contain(Telemetry.Telemetry.WelcomeMessage);
             }
             finally
@@ -373,7 +334,7 @@ namespace Microsoft.DotNet.Interactive.App.Tests.CommandLine
             Environment.SetEnvironmentVariable(environmentVariableName, "1");
             try
             {
-                await _parser.InvokeAsync(String.Empty, _console);
+                await _parser.InvokeAsync("jupyter", _console);
                 _console.Out.ToString().Should().NotContain(Telemetry.Telemetry.WelcomeMessage);
             }
             finally
